+++ src/aui-editable-deprecated/HISTORY.md
@@ -1,2 +1,7 @@
-aui-editable-deprecated
-========
+# AUI Editable Deprecated
+
+> Documentation and test modifications are not included in this changelog. For more details, see [full commit history](https://github.com/liferay/alloy-ui/commits/master-deprecated/src/aui-editable-deprecated).
+
+## @VERSION@
+
+* [AUI-1836](https://issues.liferay.com/browse/AUI-1836) Hide editable on window resize